refactor(api): move checkAuthorizationToAccessCampaign prehandler to prescription/campaign
