+++ api/src/prescription/campaign/application/security-pre-handlers.js
@@ -1,4 +1,5 @@
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
+import * as checkAuthorizationToAccessCampaignUsecase from './usecases/checkAuthorizationToAccessCampaign.js';
 import * as checkAuthorizationToManageCampaignUsecase from './usecases/checkAuthorizationToManageCampaign.js';
 import * as checkCampaignBelongsToCombinedCourseUsecase from './usecases/checkCampaignBelongsToCombinedCourse.js';
 
@@ -29,7 +30,24 @@ async function checkAuthorizationToManageCampaign(
   return securityPreHandlers.replyForbiddenError(h);
 }
 
+async function checkAuthorizationToAccessCampaign(
+  request,
+  h,
+  dependencies = { checkAuthorizationToAccessCampaignUsecase },
+) {
+  const userId = request.auth.credentials.userId;
+  const campaignId = request.params.campaignId || request.params.id;
+  const belongsToOrganization = await dependencies.checkAuthorizationToAccessCampaignUsecase.execute({
+    userId,
+    campaignId,
+  });
+
+  if (belongsToOrganization) return h.response(true);
+  return securityPreHandlers.replyForbiddenError(h);
+}
+
 export const campaignSecurityPreHandlers = {
   checkCampaignBelongsToCombinedCourse,
   checkAuthorizationToManageCampaign,
+  checkAuthorizationToAccessCampaign,
 };
