+++ api/src/shared/application/security-pre-handlers.js
@@ -16,7 +16,6 @@ import * as checkAdminMemberHasRoleCertifUseCase from './usecases/checkAdminMemb
 import * as checkAdminMemberHasRoleMetierUseCase from './usecases/checkAdminMemberHasRoleMetier.js';
 import * as checkAdminMemberHasRoleSuperAdminUseCase from './usecases/checkAdminMemberHasRoleSuperAdmin.js';
 import * as checkAdminMemberHasRoleSupportUseCase from './usecases/checkAdminMemberHasRoleSupport.js';
-import * as checkAuthorizationToAccessCampaignUsecase from './usecases/checkAuthorizationToAccessCampaign.js';
 import * as checkOrganizationDoesNotHaveFeatureUseCase from './usecases/checkOrganizationDoesNotHaveFeature.js';
 import * as checkOrganizationHasFeatureUseCase from './usecases/checkOrganizationHasFeature.js';
 import * as checkOrganizationIsNotManagingStudentsUseCase from './usecases/checkOrganizationIsNotManagingStudents.js';
@@ -548,22 +547,6 @@ async function checkUserBelongsToOrganization(request, h, dependencies = { check
   return _replyForbiddenError(h);
 }
 
-async function checkAuthorizationToAccessCampaign(
-  request,
-  h,
-  dependencies = { checkAuthorizationToAccessCampaignUsecase },
-) {
-  const userId = request.auth.credentials.userId;
-  const campaignId = request.params.campaignId || request.params.id;
-  const belongsToOrganization = await dependencies.checkAuthorizationToAccessCampaignUsecase.execute({
-    userId,
-    campaignId,
-  });
-
-  if (belongsToOrganization) return h.response(true);
-  return _replyForbiddenError(h);
-}
-
 function hasAtLeastOneAccessOf(securityChecks) {
   return async (request, h) => {
     const responses = await PromiseUtils.map(securityChecks, (securityCheck) => securityCheck(request, h));
@@ -759,7 +742,6 @@ export const securityPreHandlers = {
   checkAdminMemberHasRoleMetier,
   checkAdminMemberHasRoleSuperAdmin,
   checkAdminMemberHasRoleSupport,
-  checkAuthorizationToAccessCampaign,
   checkCertificationCenterIsNotScoManagingStudents,
   checkOrganizationHasFeature,
   checkRequestedUserIsAuthenticatedUser,
