+++ api/tests/shared/unit/application/security-pre-handlers_test.js
@@ -1336,79 +1336,6 @@ describe('Shared | Unit | Application | SecurityPreHandlers', function () {
     });
   });
 
-  describe('#checkAuthorizationToAccessCampaign', function () {
-    context('Successful case', function () {
-      it('should authorize access to resource when the user is authenticated and is admin in organization and owner of the campaign', async function () {
-        // given
-        const request = {
-          auth: { credentials: { accessToken: 'valid.access.token', userId: Symbol('UserId') } },
-          params: { id: Symbol('campaignId') },
-        };
-
-        const checkAuthorizationToAccessCampaignUsecaseStub = {
-          execute: sinon.stub(),
-        };
-        checkAuthorizationToAccessCampaignUsecaseStub.execute
-          .withArgs({ campaignId: request.params.id, userId: request.auth.credentials.userId })
-          .resolves(true);
-        // when
-        const response = await securityPreHandlers.checkAuthorizationToAccessCampaign(request, hFake, {
-          checkAuthorizationToAccessCampaignUsecase: checkAuthorizationToAccessCampaignUsecaseStub,
-        });
-
-        // then
-        expect(response.source).to.be.true;
-      });
-
-      it('should use campaignId param if id is not provided', async function () {
-        // given
-        const request = {
-          auth: { credentials: { accessToken: 'valid.access.token', userId: Symbol('UserId') } },
-          params: { campaignId: Symbol('campaignId') },
-        };
-
-        const checkAuthorizationToAccessCampaignUsecaseStub = {
-          execute: sinon.stub(),
-        };
-        checkAuthorizationToAccessCampaignUsecaseStub.execute
-          .withArgs({ campaignId: request.params.campaignId, userId: request.auth.credentials.userId })
-          .resolves(true);
-        // when
-        const response = await securityPreHandlers.checkAuthorizationToAccessCampaign(request, hFake, {
-          checkAuthorizationToAccessCampaignUsecase: checkAuthorizationToAccessCampaignUsecaseStub,
-        });
-
-        // then
-        expect(response.source).to.be.true;
-      });
-    });
-
-    context('Error cases', function () {
-      it('should forbid resource access when user is member but does not own the campaign', async function () {
-        // given
-        const request = {
-          auth: { credentials: { accessToken: 'valid.access.token', userId: Symbol('UserId') } },
-          params: { id: Symbol('campaignId') },
-        };
-
-        const checkAuthorizationToAccessCampaignUsecaseStub = {
-          execute: sinon.stub(),
-        };
-        checkAuthorizationToAccessCampaignUsecaseStub.execute
-          .withArgs({ campaignId: request.params.id, userId: request.auth.userId })
-          .resolves(false);
-        // when
-        const response = await securityPreHandlers.checkAuthorizationToAccessCampaign(request, hFake, {
-          checkAuthorizationToAccessCampaignUsecase: checkAuthorizationToAccessCampaignUsecaseStub,
-        });
-
-        // then
-        expect(response.statusCode).to.equal(403);
-        expect(response.isTakeOver).to.be.true;
-      });
-    });
-  });
-
   describe('#checkUserIsMemberOfCertificationCenterSessionFromCertificationCourseId', function () {
     context('Successful case', function () {
       it('should authorize access to resource when the user is a member of the organization center', async function () {
